Turning down the contrast.
The lines for tabs and the dots for spaces where too bright for all Brackets themes so I toned them down to have less contrast and compared it to SublimeText. If you compare the change to SublimeText you'll notice it's closer to SublimeText's contrast after this change as well. Cheers!